refactor : ui 폴더로 이동
