--- app/src/main/java/com/example/mungge_groom/home/MainActivity.kt
@@ -1,11 +1,0 @@
-package com.example.mungge_groom.home
-
-import com.example.mungge_groom.R
-import com.example.mungge_groom.databinding.ActivityMainBinding
-import com.example.mungge_groom.base.BaseActivity
-
-class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
-    override fun setLayout() {
-
-    }
-}
