+++ app/src/main/java/com/example/mungge_groom/ui/account/LoginActivity.kt
@@ -1,8 +1,8 @@
-package com.example.mungge_groom.account
+package com.example.mungge_groom.ui.account
 
-import com.example.mungge_groom.home.MainActivity
+import com.example.mungge_groom.ui.home.MainActivity
 import com.example.mungge_groom.R
-import com.example.mungge_groom.base.BaseActivity
+import com.example.mungge_groom.ui.base.BaseActivity
 import com.example.mungge_groom.databinding.ActivityLoginBinding
 
 class LoginActivity : BaseActivity<ActivityLoginBinding>(R.layout.activity_login) {
