+++ app/src/main/java/com/example/mungge_groom/ui/chat/ChatRoomFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mungge_groom.chat
+package com.example.mungge_groom.ui.chat
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
