+++ app/src/main/java/com/example/mungge_groom/ui/home/Toggle1Fragment.kt
@@ -1,4 +1,4 @@
-package com.example.mungge_groom.home
+package com.example.mungge_groom.ui.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.mungge_groom.R
-import com.example.mungge_groom.base.BaseFragment
+import com.example.mungge_groom.ui.base.BaseFragment
 import com.example.mungge_groom.databinding.FragmentToggle1Binding
