+++ app/src/main/java/com/example/mungge_groom/ui/mypage/MyPageFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mungge_groom.chat
+package com.example.mungge_groom.ui.mypage
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [ChatListFragment.newInstance] factory method to
+ * Use the [MyPageFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class ChatListFragment : Fragment() {
+class MyPageFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class ChatListFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_chat_list, container, false)
+        return inflater.inflate(R.layout.fragment_my_page, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class ChatListFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment ChatListFragment.
+         * @return A new instance of fragment MyPageFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ChatListFragment().apply {
+            MyPageFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
